# services/core_backend/openapi.yaml
openapi: 3.0.3
info:
  title: Continuum Core Backend API
  description: |
    REST API for the Continuum parametric income-protection insurance platform.
    Handles worker registration, policy lifecycle, claims, payouts, mandates,
    reserves, consent, and admin operations.
  version: 1.0.0
  contact:
    name: Continuum Team
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Token:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      summary: Register a new worker
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [worker_id, platform, upi_id, tier, password]
              properties:
                worker_id:
                  type: string
                platform:
                  type: string
                  enum: [swiggy, zomato]
                upi_id:
                  type: string
                tier:
                  type: string
                  enum: [silver, gold, platinum]
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Worker registered, JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
                  policy_eligible_from:
                    type: string
                    format: date-time
        "400":
          description: Missing fields or password too short
        "409":
          description: Worker already exists
  /auth/login:
    post:
      summary: Login with credentials
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [worker_id, password]
              properties:
                worker_id:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "401":
          description: Invalid credentials

  # ── Policies ────────────────────────────────────────────────
...